Make the smiley-editing UI bearable
The smiley-editing UI has a table that reacts to mouse hovering:
The hovered row of cells gets a different background.
In the default theme, the background is #e2eef8, which the dark
theme inherited (LemmingsDark didn't specify this background).
That made the white text in the smiley table row unreadable.

Now, the highlight color is dark but still lighter than the
unhighlit table rows.